+++ tests/unittest/test_nocasedict.py
@@ -365,7 +365,7 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError if TEST_AGAINST_DICT else AttributeError, None, True
     ),
     (
-        "Empty dict, with empty string key (not found)",
+        "Empty dict, with empty unicode string key (not found)",
         dict(
             obj=NocaseDict(),
             key='',
@@ -374,10 +374,28 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError, None, True
     ),
     (
-        "Empty dict, with non-empty key (not found)",
+        "Empty dict, with empty byte string key (not found)",
+        dict(
+            obj=NocaseDict(),
+            key=b'',
+            exp_value=None,
+        ),
+        KeyError, None, True
+    ),
+    (
+        "Empty dict, with non-empty unicode string key (not found)",
         dict(
             obj=NocaseDict(),
             key='Dog',
+            exp_value=None,
+        ),
+        KeyError, None, True
+    ),
+    (
+        "Empty dict, with non-empty byte string key (not found)",
+        dict(
+            obj=NocaseDict(),
+            key=b'Dog',
             exp_value=None,
         ),
         KeyError, None, True
@@ -394,7 +412,7 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError, None, True
     ),
     (
-        "Non-empty dict, with empty string key (not found)",
+        "Non-empty dict, with empty unicode string key (not found)",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='',
@@ -403,7 +421,17 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError, None, True
     ),
     (
-        "Non-empty dict, with non-empty non-existing key (not found)",
+        "Non-empty dict, with empty byte string key (not found)",
+        dict(
+            obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
+            key=b'',
+            exp_value=None,
+        ),
+        KeyError, None, True
+    ),
+    (
+        "Non-empty dict, with non-empty non-existing unicode string key "
+        "(not found)",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='invalid',
@@ -412,7 +440,17 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError, None, True
     ),
     (
-        "Non-empty dict, with existing key in original case",
+        "Non-empty dict, with non-empty non-existing byte string key "
+        "(not found)",
+        dict(
+            obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
+            key=b'invalid',
+            exp_value=None,
+        ),
+        KeyError, None, True
+    ),
+    (
+        "Non-empty dict, with existing unicode string key in original case",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='Dog',
@@ -421,7 +459,17 @@ TESTCASES_NOCASEDICT_GETITEM = [
         None, None, True
     ),
     (
-        "Non-empty dict, with existing key in non-original upper case",
+        "Non-empty dict, with existing byte string key in original case",
+        dict(
+            obj=NocaseDict([(b'Dog', 'Cat'), (b'Budgie', 'Fish')]),
+            key=b'Dog',
+            exp_value='Cat',
+        ),
+        None, None, True
+    ),
+    (
+        "Non-empty dict, with existing unicode string key in non-original "
+        "upper case",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='DOG',
@@ -430,7 +478,18 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError if TEST_AGAINST_DICT else None, None, True
     ),
     (
-        "Non-empty dict, with existing key in non-original lower case",
+        "Non-empty dict, with existing byte string key in non-original "
+        "upper case",
+        dict(
+            obj=NocaseDict([(b'Dog', 'Cat'), (b'Budgie', 'Fish')]),
+            key=b'DOG',
+            exp_value='Cat',
+        ),
+        KeyError if TEST_AGAINST_DICT else None, None, True
+    ),
+    (
+        "Non-empty dict, with existing unicode string key in non-original "
+        "lower case",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='dog',
@@ -439,13 +498,52 @@ TESTCASES_NOCASEDICT_GETITEM = [
         KeyError if TEST_AGAINST_DICT else None, None, True
     ),
     (
-        "Non-empty dict, with existing key in non-original mixed case",
+        "Non-empty dict, with existing byte string key in non-original "
+        "lower case",
+        dict(
+            obj=NocaseDict([(b'Dog', 'Cat'), (b'Budgie', 'Fish')]),
+            key=b'dog',
+            exp_value='Cat',
+        ),
+        KeyError if TEST_AGAINST_DICT else None, None, True
+    ),
+    (
+        "Non-empty dict, with existing unicode string key in non-original "
+        "mixed case",
         dict(
             obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
             key='doG',
             exp_value='Cat',
         ),
         KeyError if TEST_AGAINST_DICT else None, None, True
+    ),
+    (
+        "Non-empty dict, with existing byte string key in non-original "
+        "mixed case",
+        dict(
+            obj=NocaseDict([(b'Dog', 'Cat'), (b'Budgie', 'Fish')]),
+            key=b'doG',
+            exp_value='Cat',
+        ),
+        KeyError if TEST_AGAINST_DICT else None, None, True
+    ),
+    (
+        "Non-empty unicode string dict, with same byte string key",
+        dict(
+            obj=NocaseDict([('Dog', 'Cat'), ('Budgie', 'Fish')]),
+            key=b'Dog',
+            exp_value='Cat',
+        ),
+        KeyError, None, True
+    ),
+    (
+        "Non-empty byte string dict, with same unicode string key",
+        dict(
+            obj=NocaseDict([(b'Dog', 'Cat'), (b'Budgie', 'Fish')]),
+            key='Dog',
+            exp_value='Cat',
+        ),
+        KeyError, None, True
     ),
 ]